requirements.txt
asgiref==3.8.1
attrs==23.2.0
crispy-bootstrap5==2024.2
Django==5.0.6
django-allauth==0.63.3
django-appconf==1.0.6
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-imagekit==5.0.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pilkit==3.0
pillow==10.3.0
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
